services/api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.10
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via starlette
async-timeout==4.0.3
    # via aiohttp
attrs==24.2.0
    # via aiohttp
beanie==1.27.0
    # via -r requirements.in
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.7
    # via
    #   beanie
    #   uvicorn
cryptography==43.0.3
    # via python-jose
deepdiff==8.1.0
    # via -r requirements.in
dnspython==2.7.0
    # via pymongo
ecdsa==0.19.0
    # via python-jose
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.4
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
google-api-core[grpc]==2.24.2
    # via
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-auth==2.38.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-secret-manager==2.23.3
    # via -r requirements.in
google-cloud-storage==3.1.0
    # via -r requirements.in
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.69.2
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-secret-manager
grpcio==1.72.0rc1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.72.0rc1
    # via google-api-core
h11==0.14.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
    #   yarl
jsonpatch==1.33
    # via -r requirements.in
jsonpointer==3.0.0
    # via jsonpatch
lazy-model==0.2.0
    # via beanie
motor==3.6.0
    # via
    #   -r requirements.in
    #   beanie
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
numpy==1.26.4
    # via shapely
orderly-set==5.2.3
    # via deepdiff
pillow==11.0.0
    # via -r requirements.in
propcache==0.2.0
    # via yarl
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-secret-manager
protobuf==6.30.2
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   beanie
    #   fastapi
    #   lazy-model
    #   pydantic-settings
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.6.0
    # via -r requirements.in
pydash==8.0.5
    # via -r requirements.in
pyhumps==3.8.0
    # via -r requirements.in
pyjwt==2.10.1
    # via -r requirements.in
pymongo[srv]==4.9.2
    # via
    #   -r requirements.in
    #   motor
python-dotenv==1.0.1
    # via pydantic-settings
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-multipart==0.0.17
    # via -r requirements.in
pytz==2024.2
    # via -r requirements.in
pyyaml==6.0.2
    # via -r requirements.in
requests==2.32.3
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via
    #   google-auth
    #   python-jose
shapely==2.0.6
    # via -r requirements.in
six==1.16.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
starlette==0.41.2
    # via fastapi
toml==0.10.2
    # via beanie
typing-extensions==4.12.2
    # via
    #   anyio
    #   beanie
    #   fastapi
    #   multidict
    #   pydantic
    #   pydantic-core
    #   pydash
    #   uvicorn
urllib3==2.3.0
    # via requests
uvicorn==0.32.0
    # via -r requirements.in
yarl==1.17.0
    # via aiohttp
